requirements.txt
pandas>=1.3.0,<2.0
numpy>=1.21.0
matplotlib>=3.4.0
yfinance>=0.1.70
ib-insync>=0.9.17
bt>=0.2.26
tqdm>=4.62.0
pandas-datareader>=0.10.0
statsmodels>=0.13.0
pypfopt>=1.5.2
requests>=2.26.0
python-dotenv>=0.19.0
